// src/types/apps/area.ts
export interface CountryItem {
  areaCode: string
  hasNext: string
  lng: string
  name: string
  parentCode: string
}

export interface AreaItem {
  areaCode: string;
  hasNext: string;
  lng: string;
  name: string;
  parentCode: string
}

export interface CityItem {
  areaCode: string;
  hasNext: string;
  lng: string;
  name: string;
  parentCode: string
}